requirements.txt
streamlit==1.22.0
requests==2.28.1
Pillow==10.0.0 
groq==0.9.0  # Latest available version
